requirements.txt
streamlit
PyPDF2
langchain-community
langchain-openai
python-dotenv
spacy
openai
